Update Readme
* CLEAN general code;
* FIX typos;

+++ NitroCppTest-LeandroAlmeida/CustomRectangle.h
@@ -9,8 +9,8 @@ class CustomRectangle
 private:
 	class Position2D LeftTopCorner;
 	Position2D RightBottomCorner;
-	double width = 0;
-	double height = 0;
+	int width = 0;
+	int height = 0;
 
 public:
 	CustomRectangle(const Position2D x1, const Position2D x2, bool isToPrint = false);
